+++ aurral/rootfs/usr/local/bin/docker-entrypoint.sh
@@ -1,8 +1,0 @@
-#!/bin/sh
-set -e
-
-mkdir -p /config/data
-mkdir -p "${DOWNLOAD_FOLDER}"
-mkdir -p "${WEEKLY_FLOW_FOLDER}"
-
-exec "$@"