+++ webtop_kde/rootfs/etc/cont-init.d/21-gpu_permissions.sh
@@ -1,1 +1,0 @@
-../../../../claude_desktop/rootfs/etc/cont-init.d/21-gpu_permissions.sh