+++ whoogle/rootfs/etc/cont-init.d/99-run.sh
@@ -1,7 +1,0 @@
-#!/bin/sh
-
-exec misc/tor/start-tor.sh &
-./run &
-echo "Starting NGinx..."
-
-exec nginx