+++ lidarr/CHANGELOG.md
@@ -12,7 +12,7 @@
 - Update to latest version from linuxserver/docker-lidarr (changelog : https://github.com/linuxserver/docker-lidarr/releases)
 - The Home Assistant project has deprecated support for the armv7, armhf and i386 architectures. Support wil be fully dropped in the upcoming Home Assistant 2025.12 release
 
-- Added support for configuring extra environment variables via the `env_vars` add-on option alongside config.yaml. See https://github.com/alexbelgium/hassio-addons/wiki/Add-Environment-variables-to-your-Addon-2 for details.
+- Added support for configuring extra environment variables via the `env_vars` add-on option alongside config.yaml. See https://github.com/tomjansen/hassio-addons/wiki/Add-Environment-variables-to-your-Addon-2 for details.
 
 ## 3.0.1.4866 (2025-11-01)
 - Update to latest version from linuxserver/docker-lidarr (changelog : https://github.com/linuxserver/docker-lidarr/releases)